velolabs_repos/Skylock_backend/requirements.txt
alembic==0.7.7
awsebcli==3.7.6
blessed==1.9.5
blinker==1.4
botocore==1.4.19
cement==2.8.2
cffi==1.6.0
colorama==0.3.7
coverage==4.0.3
cryptography==1.1
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
enum34==1.0.4
Flask==0.10.1
Flask-Excel==0.0.3
Flask-HTTPAuth==2.5.0
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Migrate==1.4.0
Flask-MySQLdb==0.1.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Flask-WTF==0.11
future==0.15.2
gevent==1.1.0
greenlet==0.4.9
gunicorn==19.3.0
httpie==0.9.2
httplib2==0.9.1
idna==2.0
ipaddress==1.0.14
ipython==3.2.0
itsdangerous==0.24
jdcal==1.2
Jinja2==2.7.3
jmespath==0.9.0
locustio==0.7.3
Mako==1.0.1
MarkupSafe==0.23
msgpack-python==0.4.7
mysqlclient==1.3.6
openpyxl==2.2.2
parse==1.6.6
pathspec==0.3.4
phonenumbers==7.1.0
pyasn1==0.1.9
pycparser==2.14
pyexcel==0.2.1
pyexcel-io==0.1.0
pyexcel-webio==0.0.3
pyexcel-xls==0.1.0
pyexcel-xlsx==0.1.0
Pygments==2.0.2
pyotp==2.0.1
python-dateutil==2.4.2
python-gcm==0.3
pytz==2015.4
PyYAML==3.11
requests==2.7.0
semantic-version==2.5.0
six==1.9.0
SQLAlchemy==1.0.6
SQLAlchemy-Utils==0.31.2
texttable==0.8.4
twilio==4.5.0
wcwidth==0.1.6
websocket-client==0.37.0
Werkzeug==0.10.4
WTForms==2.0.2
xlrd==0.9.4
xlwt-future==0.8.0
